pta/pta/app/requirements-dev.txt
ruff
mypy
pytest
pytest-django
build
